# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: PSMS API
  version: 1.0.0
  description: |
    Placement & Service Management System — HTTP API contract.

    This file is the single source of truth for the API surface. Generate
    TypeScript types from this file with `pnpm types:generate` (the result
    is written to `packages/shared/src/types.ts`). Generate the web client
    with `pnpm api:client`.

    Every endpoint declares its required role via the `x-required-role`
    extension. Endpoints with `security: []` are public.

    Requirement traceability uses the IDs defined in `docs/specs/`:
    INP-NN inputs, PRC-NN processing, OUT-NN outputs.
  contact:
    name: Axiom AI Solutions
    email: dorien@axiom.ai
  license:
    name: Proprietary
servers:
  - url: https://psms.knox.edu.jm/api/v1
    description: Production
  - url: https://staging.psms.knox.edu.jm/api/v1
    description: Staging
  - url: http://localhost:3000/api/v1
    description: Local dev
...
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: psms_session
      description: |
        Session cookie set by `POST /auth/sign-in`. HttpOnly, Secure, SameSite=Lax.
        TTL 8 hours of inactivity. Used for Knox-authenticated roles
        (student, coordinator, admin).

    magicLinkToken:
      type: apiKey
      in: query
      name: token
      description: |
        Single-use 32-byte URL-safe token used by external supervisors.
        Issued by `POST /auth/magic-link/issue`, consumed by any endpoint
        carrying `security: [magicLinkToken: []]`. 24-hour TTL.

  # ───────────────────────────────────────────────────────────────────────
  # Reusable parameters
  # ───────────────────────────────────────────────────────────────────────
  parameters:
    pageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    pageSizeParam:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  schemas:
...
    Pagination:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    PlacementStatus:
      type: string
      enum: [PENDING_START, ACTIVE, COMPLETED, TERMINATED]
    Placement:
      type: object
      required: [placement_id, application_id, student_id, opportunity_id, status, start_date, end_date]
      properties:
        placement_id: { type: string, format: uuid }
        application_id: { type: string, format: uuid }
        student_id: { type: string, format: uuid }
        student_name: { type: string, readOnly: true }
        opportunity_id: { type: string, format: uuid }
        opportunity_title: { type: string, readOnly: true }
        organisation_name: { type: string, readOnly: true }
        supervisor_id: { type: string, format: uuid }
        supervisor_name: { type: string, readOnly: true }
        supervisor_email: { type: string, format: email, readOnly: true }
        status: { $ref: '#/components/schemas/PlacementStatus' }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        composite_rating:
          type: number
          format: float
          nullable: true
          description: "PRC-08 composite, populated after final evaluation."
    PlacementListResponse:
      allOf:
        - $ref: '#/components/schemas/Pagination'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/Placement' }
    HoursLogStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    HoursLogEntry:
      type: object
      required: [log_id, placement_id, week_number, date, hours, status]
      properties:
        log_id: { type: string, format: uuid }
        placement_id: { type: string, format: uuid }
        week_number: { type: integer, minimum: 1 }
        date: { type: string, format: date }
        hours: { type: number, format: float, minimum: 0.5, maximum: 168 }
        activity_narrative: { type: string, maxLength: 1000 }
        status: { $ref: '#/components/schemas/HoursLogStatus' }
        approved_at: { type: string, format: date-time, nullable: true }
        approved_by_supervisor_id: { type: string, format: uuid, nullable: true }
        rejection_comment: { type: string, nullable: true }
    HoursLogInput:
      type: object
      required: [week_number, date, hours, activity_narrative]
      properties:
        week_number: { type: integer, minimum: 1 }
        date: { type: string, format: date }
        hours: { type: number, format: float, minimum: 0.5, maximum: 168 }
        activity_narrative: { type: string, minLength: 10, maxLength: 1000 }
...
    SiteVisitNote:
      type: object
      required: [note_id, placement_id, visit_date, narrative, overall_assessment]
      properties:
        note_id: { type: string, format: uuid }
        placement_id: { type: string, format: uuid }
        visit_date: { type: string, format: date }
        narrative: { type: string, maxLength: 2000 }
        overall_assessment:
          type: string
          enum: [SATISFACTORY, CONCERNS, UNSATISFACTORY]
        follow_up_actions:
          type: array
          items:
            type: object
            properties:
              action: { type: string }
              due_date: { type: string, format: date }
        created_at: { type: string, format: date-time }

    # ── Evaluation (INP-06, PRC-08) ─────────────────────────────────────
...
paths:
...
  /placements:
    get:
      tags: [placements]
      summary: List placements visible to the caller
      x-required-role: ANY_AUTHENTICATED
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/pageSizeParam'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/PlacementStatus' }
      responses:
        '200':
          description: List
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PlacementListResponse' }
  /placements/{placement_id}:
    parameters:
      - name: placement_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [placements]
      x-required-role: ANY_AUTHENTICATED
      responses:
        '200':
          description: One
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Placement' }
  /placements/{placement_id}/hours:
    parameters:
      - name: placement_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [placements]
      summary: List weekly hours-log entries for a placement
      x-required-role: ANY_AUTHENTICATED
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/HoursLogEntry' }
    post:
      tags: [placements]
      summary: Submit a weekly hours log (student)
      x-required-role: STUDENT
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HoursLogInput' }
      responses:
        '201':
          description: Submitted, pending supervisor approval
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HoursLogEntry' }
...
  /placements/{placement_id}/site-visits:
    parameters:
      - name: placement_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [placements]
      summary: Record a site-visit note (coordinator) — INP-08
      x-required-role: COORDINATOR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [visit_date, narrative, overall_assessment]
              properties:
                visit_date: { type: string, format: date }
                narrative: { type: string, minLength: 10, maxLength: 2000 }
                overall_assessment:
                  type: string
                  enum: [SATISFACTORY, CONCERNS, UNSATISFACTORY]
                follow_up_actions:
                  type: array
                  items:
                    type: object
                    properties:
                      action: { type: string }
                      due_date: { type: string, format: date }
      responses:
        '201':
          description: Recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SiteVisitNote' }

  # ── Evaluations (INP-06, PRC-08) ────────────────────────────────────────
...